fix travis and appveyor error

+++ source/adios2/helper/adiosMPIFunctions.tcc
@@ -129,13 +129,14 @@ void BroadcastVector(std::vector<char> &vector, MPI_Comm mpiComm,
     }
 
     const int MAXBCASTSIZE = 1073741824;
-    int blockSize = (inputSize > MAXBCASTSIZE ? MAXBCASTSIZE : inputSize);
+    size_t blockSize = (inputSize > MAXBCASTSIZE ? MAXBCASTSIZE : inputSize);
     size_t sent = 0;
     size_t pos = 0;
     char *buffer = vector.data();
     while (inputSize > 0)
     {
-        MPI_Bcast(buffer, blockSize, MPI_CHAR, rankSource, mpiComm);
+        MPI_Bcast(buffer, static_cast<int>(blockSize), MPI_CHAR, rankSource,
+                  mpiComm);
         buffer += blockSize;
         inputSize -= blockSize;
         blockSize = (inputSize > MAXBCASTSIZE ? MAXBCASTSIZE : inputSize);
